variants are more robust now
The variations used to be an a one lists of some variant subsets. Now each variant type has fuller subtypes which are detailed out. There are two json dictionaries one for high level and another is granular. The markeddown pages of the app and github readme are up to date to reflect these changes.